OSProcess 4.3.16 Updates from Ross Boylan for Cuis support (thanks!)
Use #newProcess and #resume rather than #fork when the identity of the new process is answered. The #fork method in Cuis does not answer the process (to guard against concurrency bugs).

Fix OSProcessAccessor class>>forThisOSProcess to guard against nil ThisOSProcess.

Do not use #ifNotNilDo: (for Cuis compatibility) and do not use #ifNotNil: with a block parameter (for Squeak 3.6 compatibility).