OSProcess 4.4.4
Fix #isStdIn: #isStdOut: and #isStdErr: which should compare only the session ID an file handle portion of the fileID.

Add #testDupTo #testUnixFileNumber and #testRedirectStdOutTo to exercise dup2() and file stream redirection.

Note: Cog VMs using the oscog branch of OSProcessPlugin will fail testRedirectStdOutTo due to an error in the plugin.