OSProcess 4.4.10
Better exit status decoding for unix processes. The system call wait() answers an integer value, and the C runtime library has macros to decode its bitmap value. A unix shell relies of the decoded value, either a normal exit status or a signal number of signal causing exit.

Add class UnixProcessExitStatus to decode unix process exit status in a manner similar to that of a unix shell for GNU unix systems, and use this for ExternalUnixOSProcess>>printOn:. This may not be portable across unix systems but is nevertheless used to provide a better printString description of process exit status for external unix processes.