OSProcess 4.5.13
On Windows, do not attempt to write to stderr if it is nil, as will be the case if the Windows file handle registry has been compiled into the FilePlugin. That handle registry explicitly prevents attaching to stdio streams. A possible workaround would be to provide an alternative FilePlugin dll (external) with the handle registry disabled.