OSProcess 4.6.2
Implement newPid to clean the allMyChildren registry after a fork, stopping the child watcher first, and restarting it for the emptied registry.

Do not call updateAllMyChildren in resetChildProcessDictionary, because this may be a child of another Squeak, in which case the dictionary should be emptied without effect on the registered children of the parent Squeak.

Move grimReaper up to OSProcessAccessor.

Remove hasTestEof: