OSProcess 4.6.7
Add nil checks in #shutDown: to prevent failures when exiting existing image with a VM that does not include OSPP (e.g. SqueakJS).

For AIO event hander, if the AIO plugin is not present, write an error to console rather than interrupting the image with a notiifer. Reduces the annoyance level if VM provider provides OSPP but neglects to include the AIO plugin.